14 add custom exception

+++ HR.LeaveManagement.Application/Features/LeaveAllocations/Handlers/Commands/DeleteLeaveAllocationCommandHandler.cs
@@ -7,10 +7,12 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Threading;
 using HR.LeaveManagement.Application.Features.LeaveAllocations.Requests.Commands;
+using HR.LeaveManagement.Application.Exceptions;
+using HR.LeaveManagement.Domain.Entity;
 
 namespace HR.LeaveManagement.Application.Features.LeaveAllocations.Handlers.Commands
 {
-    public  class DeleteLeaveAllocationCommandHandler : IRequestHandler<DeleteLeaveAllocationCommand>
+    public class DeleteLeaveAllocationCommandHandler : IRequestHandler<DeleteLeaveAllocationCommand>
     {
         private readonly ILeaveAllocationRepository _leaveAllocationRepository;
         public DeleteLeaveAllocationCommandHandler(ILeaveAllocationRepository leaveAllocationRepository)
@@ -21,6 +23,10 @@ namespace HR.LeaveManagement.Application.Features.LeaveAllocations.Handlers.Comm
         public async Task<Unit> Handle(DeleteLeaveAllocationCommand request, CancellationToken cancellationToken)
         {
             var leaveAllocation = await _leaveAllocationRepository.Get(request.Id);
+
+            if (leaveAllocation == null)
+                throw new NotFoundException(nameof(HR.LeaveManagement.Domain.Entity.LeaveAllocation), request.Id);
+
             await _leaveAllocationRepository.Delete(leaveAllocation);
             return Unit.Value;
         }
