+++ HR.LeaveManagement.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -7,6 +7,8 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Threading;
 using HR.LeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
+using HR.LeaveManagement.Application.Exceptions;
+using HR.LeaveManagement.Domain.Entity;
 
 namespace HR.LeaveManagement.Application.Features.LeaveRequests.Handlers.Commands
 {
@@ -21,6 +23,10 @@ namespace HR.LeaveManagement.Application.Features.LeaveRequests.Handlers.Command
         public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
         {
             var leaveReques = await _leaveRequestRepository.Get(request.Id);
+
+            if (leaveReques == null)
+                throw new NotFoundException(nameof(LeaveRequest),request.Id);
+ 
             await _leaveRequestRepository.Delete(leaveReques);
             return Unit.Value;
         }
